adapt method moved to generic wrapper
